requirements.txt
langchain
langchain-openai
langchain-community
chromadb
torch
numpy
pandas
python-dotenv
tiktoken
